+++ ServerData.cs
@@ -143,5 +143,5 @@ public record RouteInfo(string Route, string Method, bool IsLegacy = false);
 
 public static class APIVersion
 {
-    public const string Version = "1.6.6";
+    public const string Version = "1.7.0";
 }